Split scene flow presentation
